Revenue changed from Int to Double; Tests implemented

+++ AutocompleteClient/FW/Logic/Worker/ConstructorIO.swift
@@ -44,7 +44,6 @@ public class ConstructorIO: CIOSessionManagerDelegate {
         }
     }
 
-    
     public init(config: ConstructorIOConfig) {
         self.config = config
     
@@ -140,7 +139,7 @@ public class ConstructorIO: CIOSessionManagerDelegate {
     ///   - searchTerm: Search term that the user searched for. If nil is passed, 'TERM_UNKNOWN' will be sent to the server.
     ///   - sectionName The name of the autocomplete section the term came from
     ///   - completionHandler: The callback to execute on completion.
-    public func trackConversion(itemID: String, revenue: Int?, searchTerm: String?, sectionName: String?, completionHandler: TrackingCompletionHandler? = nil){
+    public func trackConversion(itemID: String, revenue: Double?, searchTerm: String?, sectionName: String?, completionHandler: TrackingCompletionHandler? = nil){
         let data = CIOTrackConversionData(searchTerm: (searchTerm ?? "TERM_UNKNOWN"), itemID: itemID, sectionName: (sectionName ?? self.defaultItemSectionName), revenue: revenue)
         let request = self.buildRequest(data: data)
         execute(request, completionHandler: completionHandler)
